requirements.txt
click==7.1.2
cmake==3.18.4.post1
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
gcloud==0.17.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
imutils==0.5.4
jws==0.1.3
numpy==1.20.0
oauth2client==3.0.0
opencv-python==4.5.1.48
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7
six==1.15.0
